Adding some getters for tts and standardizing the app bar at the top and bottom of screen
