+++ lib/RouteGenerator.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:summer2022/enough_mail_sample.dart';
-import './mail_widget.dart';
-import './main_menu.dart';
-import './other_mail.dart';
-import './settings.dart';
-import './sign_in.dart';
+import 'ui/mail_widget.dart';
+import 'ui/main_menu.dart';
+import 'ui/other_mail.dart';
+import 'ui/settings.dart';
+import 'ui/sign_in.dart';
 import './backend_testing.dart';
 import 'models/Arguments.dart';
 import 'models/EmailArguments.dart';
@@ -18,11 +18,15 @@ class RouteGenerator {
       case '/settings':
         return MaterialPageRoute(builder: (_) => const SettingsWidget());
       case '/digest_mail':
-        return MaterialPageRoute(builder: (_) => MailWidget(digest: (settings.arguments as MailWidgetArguments).digest));
+        return MaterialPageRoute(
+            builder: (_) => MailWidget(
+                digest: (settings.arguments as MailWidgetArguments).digest));
       case '/sign_in':
         return MaterialPageRoute(builder: (_) => const SignInWidget());
       case '/other_mail':
-        return MaterialPageRoute(builder: (_) => OtherMailWidget(emails: (settings.arguments as EmailWidgetArguments).emails));
+        return MaterialPageRoute(
+            builder: (_) => OtherMailWidget(
+                emails: (settings.arguments as EmailWidgetArguments).emails));
       case '/backend_testing':
         return MaterialPageRoute(
             builder: (_) => BackendPage(title: "Backend Testing"));
